requirements.txt
asgiref==3.4.1
Django==4.0
sqlparse==0.4.2
tzdata==2021.5
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2022.1


